Added the option to remove your rating and did some minor backend improvements

+++ CrockPot/Controllers/RatingsController.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Authorization;
 using CrockPot.Models;
 using CrockPot.Services.IServices;
 using System.Security.Claims;
-using System.Diagnostics;
 
 namespace CrockPot.Controllers
 {
@@ -17,27 +16,6 @@ namespace CrockPot.Controllers
             _ratingService = ratingService;
         }
 
-        public async Task<IActionResult> Index()
-        {
-            var ratings = await _ratingService.GetRatingsAsync();
-            return View(ratings);
-        }
-
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var rating = await _ratingService.GetRatingByIdAsync(id.Value);
-            if (rating == null)
-            {
-                return NotFound();
-            }
-
-            return View(rating);
-        }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -47,44 +25,44 @@ namespace CrockPot.Controllers
 
             if (ModelState.IsValid)
             {
-                await _ratingService.CreateRatingAsync(rating);
+                var result = await _ratingService.CreateRatingAsync(rating);
+                if (!result)
+                {
+                    ModelState.AddModelError(string.Empty, "Failed to create the rating. Please try again.");
+                }
             }
 
             return RedirectToAction("Details", "Recipes", new { id = rating.RecipeId });
         }
 
-        public async Task<IActionResult> Delete(int? id)
+
+        [HttpPost, ActionName("Delete")]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Delete(int id)
         {
-            if (id == null)
+            if(id == null)
             {
                 return NotFound();
             }
+            var rating = await _ratingService.GetRatingByIdAsync(id);
 
-            var rating = await _ratingService.GetRatingByIdAsync(id.Value);
             if (rating == null)
             {
                 return NotFound();
             }
 
-            return View(rating);
-        }
-
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var success = await _ratingService.DeleteRatingAsync(id);
-            if (!success)
+            if (User.FindFirstValue(ClaimTypes.NameIdentifier) != rating.AuthorId)
+            {
+                return StatusCode(403);
+            }
+            var result = await _ratingService.DeleteRatingAsync(id);
+            if (!result)
             {
                 return Problem("Failed to delete the rating.");
             }
 
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("Details", "Recipes", new { id = rating.RecipeId });
         }
 
-        private bool RatingExists(int id)
-        {
-            return _ratingService.RatingExists(id);
-        }
     }
 }
